rename and fix import for shadcn utils and constants
